Update stats.py and clean repository

+++ src/stats.py
@@ -5,11 +5,16 @@ from google.cloud import storage
 from q1_memory import q1_memory
 from q2_memory import q2_memory
 from q3_memory import q3_memory
+from q1_time import q1_time
+from q2_time import q2_time
+from q3_time import q3_time
+import pandas as pd
 import time
 import io
 import os
 import tempfile
 
+#os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "../sa.json"
 
 def get_file_path(bucket_name, file_name):
     
@@ -18,7 +23,7 @@ def get_file_path(bucket_name, file_name):
     bucket = storage_client.bucket(bucket_name)
     file_blob = bucket.blob(file_name)
 
-     # Create a temporary file
+    # Create a temporary file
     _, temp_local_filename = tempfile.mkstemp()
 
     # Download the file to the temporary file
@@ -39,7 +44,7 @@ def memory_update_csv_in_gcs(function, bucket_name, file_name, mem_name):
     avg_memory = sum(mem_usage) / len(mem_usage)
     max_memory = max(mem_usage)
 
-    mem_row = [function.__name__, datetime.now().isoformat(), f"{avg_memory:.4f}", f"{max_memory:.4f}"]
+    mem_row = [function.__name__, int(datetime.utcnow().timestamp()), f"{avg_memory:.4f}", f"{max_memory:.4f}"]
     blob = bucket.blob(mem_name)
 
     if blob.exists():
@@ -47,9 +52,15 @@ def memory_update_csv_in_gcs(function, bucket_name, file_name, mem_name):
         csv_data = io.StringIO(existing_csv)
         csv_reader = csv.reader(csv_data)
         rows = list(csv_reader)
+        
+        df = pd.DataFrame(rows[1:], columns=rows[0])
+        filtered_df = df[df['function'] == function.__name__].sort_values(by='timestamp').iloc[-1]
+        
+        if float(filtered_df['avg_memory_usage']) <= avg_memory and float(filtered_df['max_memory_usage']) <= max_memory: return False
+        
         blob.delete()
     else:
-        rows = [['function', 'datetime', 'avg_memory_usage', 'max_memory_usage']]
+        rows = [['function', 'timestamp', 'avg_memory_usage', 'max_memory_usage']]
 
     # Add new row
     rows.append(mem_row)
@@ -65,6 +76,8 @@ def memory_update_csv_in_gcs(function, bucket_name, file_name, mem_name):
     print(f"CSV updated and uploaded to gs://{bucket_name}/{mem_name}")
     print(mem_row)
 
+    return True
+
 def efficiency_update_csv_in_gcs(function, bucket_name, file_name, eff_name):
     
     temp_file_path = get_file_path(bucket_name, file_name)
@@ -76,7 +89,7 @@ def efficiency_update_csv_in_gcs(function, bucket_name, file_name, eff_name):
     mem_usage = memory_usage((function, (temp_file_path,), {}))
     total_time = time.time() - start_time
 
-    eff_row = [function.__name__, datetime.now().isoformat(), f"{total_time:.4f}"]
+    eff_row = [function.__name__, int(datetime.utcnow().timestamp()), f"{total_time:.4f}"]
     blob = bucket.blob(eff_name)
 
     if blob.exists():
@@ -84,9 +97,15 @@ def efficiency_update_csv_in_gcs(function, bucket_name, file_name, eff_name):
         csv_data = io.StringIO(existing_csv)
         csv_reader = csv.reader(csv_data)
         rows = list(csv_reader)
+
+        df = pd.DataFrame(rows[1:], columns=rows[0])
+        filtered_df = df[df['function'] == function.__name__].sort_values(by='timestamp').iloc[-1]
+        
+        if float(filtered_df['total_time']) <= total_time: return False
+        
         blob.delete()
     else:
-        rows = [['function', 'datetime', 'total_time']]
+        rows = [['function', 'timestamp', 'total_time']]
 
     # Add new row
     rows.append(eff_row)
@@ -102,11 +121,24 @@ def efficiency_update_csv_in_gcs(function, bucket_name, file_name, eff_name):
     print(f"CSV updated and uploaded to gs://{bucket_name}/{eff_name}")
     print(eff_row)
 
+    return True
+
 if __name__ == '__main__':
     bucket_name = 'test_jobs144'
     file_name = 'lt/farmers-protest-tweets-2021-2-4.json'
+    #file_name = 'lt/twitter_test.json'
     mem_name = 'lt/results/memory_result.csv'
     eff_name = 'lt/results/efficiency_result.csv'
 
-    memory_update_csv_in_gcs(q1_memory, bucket_name, file_name, mem_name)
-    efficiency_update_csv_in_gcs(q1_memory, bucket_name, file_name, eff_name)
+    is_q1_mem = memory_update_csv_in_gcs(q1_memory, bucket_name, file_name, mem_name)
+    is_q2_mem = memory_update_csv_in_gcs(q2_memory, bucket_name, file_name, mem_name)
+    is_q3_mem = memory_update_csv_in_gcs(q3_memory, bucket_name, file_name, mem_name)
+    
+    is_q1_eff = efficiency_update_csv_in_gcs(q1_time, bucket_name, file_name, eff_name)
+    is_q2_eff = efficiency_update_csv_in_gcs(q2_time, bucket_name, file_name, eff_name)
+    is_q3_eff = efficiency_update_csv_in_gcs(q3_time, bucket_name, file_name, eff_name)
+
+    bool_list = [is_q1_mem, is_q2_mem, is_q3_mem,
+                is_q1_eff, is_q2_mem, is_q3_eff]
+
+    assert not all(value is False for value in bool_list), "AssertionError: Optimization metrics should improve not decrease."
